docker/build-model/requirements.txt
torch
torchvision
numpy
huggingface_hub
python-docx
fastapi
uvicorn
sentence-transformers
faiss-cpu
datasets
transformers